Refactored LaTeXML::Post into a master to manage post processing (mainly ProcessChain), and also export more useful forms of error messages (mimicking the API used during conversion). (Fatal,Error,Warn,Info,NoteBegin,NoteEnd,NoteProgress)
There's a new LaTeXML::Post::Processor which is the abstract
base class of all post-processing modules.

This gives a more uniform & useful set of error messages
and makes it easier to classify & detect processing
problems in batch contexts.

It probably will lead to some other nice refactoring
in the future... watch this space!